+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- `schema-viz.core/visualize-schemas` as the new public api.
+
 ## 0.0.1 (10.3.2016)
 
 Initial version.